Separate Watch and Write streams into their own files; delete FSTStream leftovers
